requirements.txt
pillow==10.2.0
opencv-python==4.9.0.80
pyarrow==15.0.0